requirements.txt
groq
ollama
python-dotenv
spacy
streamlit
pdfplumber
python-docx
reportlab
langdetect
openai
deep-translator
